# [PERF] purchase_stock:16.0 improve  get related invoices kasm