@robodoo 2024-09-27
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/181822.png)](https://mergebot.odoo.com/odoo/odoo/pull/181822)