@Whenrow 2024-10-25
@MohamedKasem99 I'm ok with the fix. We can ask @odoo/rd-runbot to check the new fields is ok to bypass the upgrade error :)