@Williambraecky 2024-10-29
Hi @MohamedKasem99 

we are a bit suspicious of the two-ways _computed_ many2many field.
Could we perhaps revert back to the [previous version of the fix](https://github.com/odoo/odoo/commit/b021278821f2be3bfa957280e54ae553a6f2b894)?

I don't think the model changes are strictly required for this performance fix.
Feel free to add the field in master though.