@MohamedKasem99 2024-11-04
@Williambraecky 

Actually after thinking more about this, I think the first version of the fix was working fine. I thought it was buggy because of the way relations are described between purchase.order and stock.picking models. I will revert back to this fix.