@robodoo 2024-11-06
@mohamedkasem99 @Whenrow staging failed: ci/runbot (view more at https://runbot.odoo.com/runbot/build/70284363)